requirements.txt
openai
tiktoken
requests
dotenv
flask
psycopg2-binary